Move MetaSolverHelper class to testing
