+++ src/test/java/edu/kit/provideq/toolbox/api/SatSolverTest.java
@@ -3,13 +3,12 @@ package edu.kit.provideq.toolbox.api;
 import static edu.kit.provideq.toolbox.SolutionStatus.SOLVED;
 import static org.hamcrest.Matchers.is;
 
-import com.google.common.collect.Lists;
 import edu.kit.provideq.toolbox.GamsProcessRunner;
 import edu.kit.provideq.toolbox.MetaSolverProvider;
 import edu.kit.provideq.toolbox.ResourceProvider;
 import edu.kit.provideq.toolbox.Solution;
 import edu.kit.provideq.toolbox.SubRoutinePool;
-import edu.kit.provideq.toolbox.meta.MetaSolverHelper;
+import edu.kit.provideq.toolbox.MetaSolverHelper;
 import edu.kit.provideq.toolbox.sat.MetaSolverSat;
 import edu.kit.provideq.toolbox.sat.SolveSatRequest;
 import edu.kit.provideq.toolbox.sat.solvers.GamsSatSolver;
